Return Position[] from book route
